shield-wall/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-multipart==0.0.18
websockets==14.*
pydantic==2.10.*
pydantic-settings==2.*
openai==1.*
google-genai==0.3.*
python-dotenv==1.0.*
aiofiles==24.*
pandas==2.2.*
numpy==2.1.*
chromadb==0.5.*
openpyxl==3.1.*
python-docx==1.1.*
PyPDF2==3.0.*
tiktoken==0.8.*
